learning: %3F

+++ src/3.c
@@ -1,11 +1,11 @@
-/* /*, =, while(x <=>= y) {asdfasd}, *, +, -, %d, x */
+/* /*, =, while(x <=>= y) {asdfasd}, *, +, -, %d, x, float */
 #include <stdio.h>
 /* print Fahrenheit-celsius table 
 for fahr = 0, 20, ..., 300 */
 int main()
 {
-    int fahr, celsius;
-    int lower, upper, step;
+    float fahr, celsius;
+    float lower, upper, step;
 
     lower = 0;     /* lower limit of temperature scale */
     upper = 300;     /* upper limit */
@@ -14,7 +14,7 @@ int main()
     fahr = lower;
     while (fahr <= upper) {
         celsius = 5 * (fahr-32) / 9;
-        printf("%d\t%d\n", fahr, celsius);
+        printf("%f\t%f\n", fahr, celsius);
         fahr = fahr + step;
     }
 }
